requirements.txt
sqlmodel>=0.0.8
matplotlib>=3.5.0
python-dotenv>=0.19.0